enable custom templates for honeybee core modules
This change makes it possible to put custom localized templates into the
`project.module_templates_dir` that is located in the
`project/app/module_templates` directory of the honeybee project. For
this the template lookup paths are adjusted in the `output_types.xml`
and AgaviConfig settings for paths are defined in `config.php`.

Another fact is the possibility to have a project specific
`translation.xml` file in `project/app/config/translation.xml`.

Additionally the upper left `Honeybee` brand name in the header template
can now be changed via translations.