removed `composer.lock` and added it to .gitignore
The `composer.lock` was always modified in the submodule which leads to
some confusion and additional steps (like git stash) when working with
honeybee directly. As we use pretty much hardcoded version numbers in
the `composer.json` file anyways it should pose no problem to use to
have the composer.lock on ignore and not under version control. If this
turns out to be wrong, we can always resubmit the file. :-)