Tree view now defaults to 10000 documents max
The treeview uses the list view to get all items and then displays those
documents as a nested tree. It used the defaultLimit of the module that
acts as a tree to get items instead of the hard limit of 10000 items as
that AgaviRequestDataHolder instance was incorrectly created and thus
lead to the fallback of 20 defaultLimit items being fetched. This meant
that only those 20 items had correctly fetched data and were thus
displayed and rendered correctly in the frontend. This is now fixed.
refs #12412 and refs #12417 as batch delete now works correctly again as
well.