Add actual validation to list state fields
The list state validator that created validators for limit, offset and
other fields necessary for a correct filter query and pagination did not
actually validate the limit and offset values to be positive integer
values. This is fixed now by validating a valid range of 0 - 2147483647.
The field sorting now is not case insensitive anymore and the search
field expects at least one character of data. What is still missing at
the moment is validation error messages that are a bit more expressive.
This refs #12399.