ReferenceFilter now resolves asset fields.
When using the reference filter for exports it is now possible to refer to 'duck-typed' binary fields which will be resolved to a set of meta data such as size and mime-type for example. The latter duck-typing is achieved by setting an 'acts_as_assets_field' option on the integer-collection fields that are used to hold asset ids.