Extracted 'Config' namespace from Import & Export.
Previously both the Export and the Import packages where using their own Config implementation.
First there was the Import which had a Config package and then came the Export and copy&pasted it.
At first it looked as though this could make sense due to the different context of the Config would operate in (Import or Export).
Turns out that a common config package is far better as the IConfig is very small and generic (get, has, set).
So we now have a Honeybee\Core\Config package.
The Export and the Import modules have been adjusted to respect the new Config location accordingly.