modified default list renderer output
The default list renderer just outputs the value. This commit changes this
behaviour to something more appropriate for the time being. The value is
being stripped of tags and shortend to 30 words if the value is longer than
255 characters. If the value is a array it will be imploded with commas.
This should be configurable, but as this is a temporary solution for the
Common module (which is about to change or being removed) it is okay atm.