improve auto TextExcerptFilter by various degrees
The filter is more stable now. It now supports excerpting by character,
word, sentence or paragraph count while being able to skip the first N
occurrences of the chosen cut method.

There are multiple settings that may be of use. They are set to sensible
defaults for auto excerpting texts so one usually only needs to specify
the number of characters, words, sentences or paragraphs one wants.
- characters: maximum number of characters (default: 250)
- words: maximum number of words
- sentences: maximum number of sentences (ending with ".", "?" or "!")
- paragraphs: maximum number of paragraphs (actual html "</p>"s)
- skip: number of characters|words|sentences|paragraphs to skip at the
        beginning (strip from start)
- ellipsis: string to use as an ellipsis (default is: " […]")
- append_ellipsis: whether to append the ellipsis or not (default: true)
- strip_tags: whether to strip html tags before excerpting (default: true)
- allowed_tags: tags not to strip if "strip_tags" is used (format: "<a><p>")
- strip_newlines: whether to strip "\n", "\r" and "\r\n" (default: true)
- strip_tabs: whether to strip tab characters (default: true)
- strip_excessive_whitespace: strip duplicate whitespaces? (default: true)

A mix of cut methods is not supported (e.g. N words while keeping a maximum
of M characters).

For more details see the source file.