intergrated the tree-sidebar component into core.
you can switch the sidebar feature on and off with the {module}.sidebar.trees.enabled setting.
the xml syntax for defining itemAction within a module's list config (settings.xml) has changed.
plan is to migrate the syntax for batch accordingly in near future. look at the module build template for examples.
reference batch dialogs are now natively supported for all reference fields of a module.
there is no more up to it than activating it for a field via list config batch action. again see the build template for examples.
the batch action dropdown is now context sensitive to the current selection of items, so it will only enable actions that are executable for all selected items.
the drag 'n drop and click reference batch assignments provided by the sidebar are now also aware of the current selection and constrain actions accordingly.