sidebar now adjusts height to list view height
The sidebar now adjusts it's height to the height of the table list view
or the viewport height in case of few entries in the list.

The factories.xml file now has some default parameters to adjust if
someone searches for them. Saves a bit of Agavi parameter names knowledge.