Introduced a new user module.
Migrated the current Auth actions to the new User module and eliminated the Auth module.
Introduced two new actions: ResetPassword and SetPassword, which introduce a new honeybee dependency on sendmail.
The ResetPasssword action is used to initiate the SetPassword action by generating a secure token and sending to the related user.
This token can be used to change the password within 20minutes.