remove font-awesome dependency and add own font
The font-awesome dependency is replaced with our own icon font from now
on. The icons can be added via css class (hb-icon-*) or via data-hb-icon
attribute. The icon-* classes have been replaced with new icons. These may
not be final, but are a first step in removing twitter bootstrap
dependencies.