updated kineticjs and adjusted ichiejs as required
with firefox 20.0 the kinecticjs Layer.visible option causes an js runtime error when passed in as a default.
error: Value '[object CanvasRenderingContext2D]' is not a valid value for enumeration CanvasWindingRule
this does not happen in safari, chrome, opera and ie 8-10.
by not passing in the default anymore we can avoid this in an acceptable way luckily.
when there is time, we might want to inspect this and file a firefox bug if it turns out to be one.