First logging system changes & Monolog integration
The AgaviLoggerManager and the AgaviLogger are now replaced by our own
LoggerManager and Logger. In addition with the MonologLoggerAppender being
introduced we have now the possibility to use the Agavi logging.xml stuff
for libs that want to use a PSR-3 compatible logger. This is possible by
using the Psr3Logger that wraps an AgaviLogger. Our logger can provide
such an instance by calling `getPsr3Logger()`.

Monolog as a logging library and Psr\Log are new dependencies. This
enables us to use the Monolog handlers and processors by creating a setup
class that configures a Monolog logger instance and using that through the
MonologLoggerAppender that has been introduced with this commit.

As an example there is a DefaultSetup class that create a Monolog logger
that logs all current log messages above DEBUG level ONLY when a CRITICAL
log message occurs. The logger logs to a file and the syslog. This
functionality can be used to create e.g. a logger appender that only sends
mails when a critical error occurs but that contains all log messages from
that request.

There is a new Agavi patch to introduce more log levels to Agavi. These
log levels basically map the Syslog standard and add TRACE to those
levels. This patch makes it easier to map Agavi log levels to PSR-3
compatible log levels without changing the Agavi logging system to be
PSR-3 compatible (and thus not backwards compatible for PHP <5.3 etc.). I
hope the redifinition of certain constant values does not break anything.

The BaseAction and BaseView now contain convenience methods like logInfo()
etc. The logger name is defined by the `getLoggerName()` method in the
base view and action and may be overridden by specific actions and views.
This makes it possible for certain custom actions to define their own
loggers and use them by overwriting `getDefaultLogger()` with that name
and still using the `$this->logDebug()` etc. convenience methods.