Add MailTemplateService and TemplateService
The MailTemplateService uses the TemplateService to render Twig templates
using the Honeybee TwigRenderer from output_types.xml. The template
service uses a 'template' output type that was newly defined. The usage of
output types was preferred for the moment to have only one place to define
the lookup path of templates (the sandbox content layer definition of
targets). It may be changed to mail.xml settings or be working without
Agavi later on.