added environaut to honeybee.
honeybee serves a pool of default checks that reflect the minimal system
requirements. projects are required to provide a specific environaut
configuration that must at least refer to honeybee's default
specification. further requirements that evolve from the project can
then easily be added on top, thereby allowing to alter the default
behaviour if required.
the new 'check-env' make target replaces the execution of the current
configure script.
the previously used bin/configure script is deprecated from now on and
will be removed soon.