Login refactoring. Users will now be redirected
correctly in more cases. HTTP 303 and 302 redirects after login. The HTTP
REFERER will be used when it matches the current base href and the method
is GET to redirect the user to the page that had the POST html form in the
first place. Internal forwarding to Login works as it was before with
transparent internal forward to the original action.

The REFERER is validated using the Symfony/Validator component as a test.
It checks for valid URLs that start with the same base href the
application has.