allow marker to be set w/o changing address fields
When the marker is moved the lat/lon values change, but the other address
field values don't change. This allows to place the marker at the entrance
of buildings etc. without changing the official address.

There is a new 'reverse_geocode' option to be set on location fields
that's propagated to the location widget. This allows for enabling of
reverse geocoding on the client after changing the location marker's
position.

refs #20679