added manual coordinates input on location widget
It's now possible to configure whether the lat/lon coordinate input fields
should be shown to users. There's a new 'visible_coordinates' option on
the location input renderer that is used to show or hide the input fields.

Upon changing the input field values of either the latitude or longitude
fields the marker position is changed.

Additionally the marker position via map clicking is now rounded to 6
decimal places and will be even shorter when the trailing digits are 0.

refs #20695